+++ .storybook/config.js
@@ -22,7 +22,7 @@ import Navigation from '../src/components/navigation';
 import ExperienceCard from '../src/components/cards/experience-card';
 import PartnerCard from '../src/components/cards/partner-card';
 import ShortExperiencePreview from '../src/components/experience/short-experience-preview/ShortExperiencePreview';
-import ExperienceListingCard from '../src/components/cards/experience-listing-card';
+import ExperienceListingCard from '../src/components/experience/experience-listing-card';
 
 // Register custom components.
 Vue.component('avatar', Avatar);
